Reshape folder studio bench flow
